+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Button theme structure changed
+- Dialog custom transitions, overlay color, opacity. Props and theme structure changed.
 
 ## [1.0.0-alpha.5] - 2019-07-29 [YANKED]